Move validation out of `apis/`.
